requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
torch
torchvision
Pillow
requests
numpy<2.0
opencv-python-headless
tqdm